pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "pydantic-fixturegen"
version = "1.1.0"
description = "Deterministic fixture generator for Pydantic models."
requires-python = ">=3.10,<3.15"
readme = "README.md"
authors = [{ name = "Fixturegen Developers" }]
license = { text = "MIT" }
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "License :: OSI Approved :: MIT License",
  "Development Status :: 5 - Production/Stable",
]
dependencies = [
  "faker>=3.0.0",
  "pydantic>=2.12.4",
  "typer>=0.12.4,<0.13",
  "click>=8.1.7,<8.2",
  "pluggy>=1.5.0",
  "tomli>=2.0.1; python_version < '3.11'",
]

[project.optional-dependencies]
regex = [
  "rstr>=3.2.2",
]
fastapi = [
  "fastapi>=0.115.0",
  "uvicorn>=0.32.0",
]
openapi = [
  "datamodel-code-generator>=0.35.0,<0.36.0",
  "pyyaml>=6.0.1",
]
email = [
  "email-validator>=2.0.0",
]
payment = [
  "pydantic-extra-types>=2.6.0",
  "python-ulid>=2.2.0",
]
orjson = [
  "orjson>=3.11.1",
]
hypothesis = [
  "hypothesis>=1.0.0",
]
dev = [
  "mypy>=1.11",
]
lint = [
  "ruff>=0.6.5",
]
test = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
]
numpy = [
  "numpy>=2.2.6; python_version < '3.11'",
  "numpy>=2.3.2; python_version >= '3.11'",
]
watch = [
  "watchfiles>=0.20.0",
]
docs = []
all = [
  "fastapi>=0.115.0",
  "uvicorn>=0.32.0",
  "datamodel-code-generator>=0.35.0,<0.36.0",
  "pyyaml>=6.0.1",
  "email-validator>=2.0.0",
  "pydantic-extra-types>=2.6.0",
  "python-ulid>=2.2.0",
  "rstr>=3.2.2",
  "orjson>=3.11.1",
  "hypothesis>=1.0.0",
  "numpy>=2.2.6; python_version < '3.11'",
  "numpy>=2.3.2; python_version >= '3.11'",
  "watchfiles>=0.20.0",
]
all-dev = [
  "mypy>=1.11",
  "ruff>=0.6.5",
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.32.0",
  "datamodel-code-generator>=0.35.0,<0.36.0",
  "pyyaml>=6.0.1",
  "email-validator>=2.0.0",
  "pydantic-extra-types>=2.6.0",
  "python-ulid>=2.2.0",
  "rstr>=3.2.2",
  "orjson>=3.11.1",
  "hypothesis>=1.0.0",
  "numpy>=2.2.6; python_version < '3.11'",
  "numpy>=2.3.2; python_version >= '3.11'",
  "watchfiles>=0.20.0",
]

[project.urls]
Homepage = "https://github.com/casper-kristiansson/pydantic-fixturegen"
Documentation = "https://github.com/casper-kristiansson/pydantic-fixturegen"
Repository = "https://github.com/casper-kristiansson/pydantic-fixturegen"

[project.scripts]
pfg = "pydantic_fixturegen.cli:app"

[project.entry-points."pydantic_fixturegen"]
builtin = "pydantic_fixturegen.plugins.builtin"

[project.entry-points."pytest11"]
pydantic_fixturegen = "pydantic_fixturegen.testing.pytest_plugin"

[tool.pytest.ini_options]
addopts = [
  "-q",
  "--strict-config",
  "--strict-markers",
  "--maxfail=1",
  "-Werror",
]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["pydantic_fixturegen"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
]
fail_under = 90
precision = 2
show_missing = true
skip_empty = true

[tool.mypy]
python_version = "3.10"
strict = true
warn_unreachable = true
warn_unused_configs = true
namespace_packages = true

[[tool.mypy.overrides]]
module = "typer"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "typer.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "orjson"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "faker"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "faker.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rstr"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "numpy"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "numpy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pydantic_extra_types"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pydantic_extra_types.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "datamodel_code_generator"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "datamodel_code_generator.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "fastapi"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "fastapi.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "uvicorn"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "uvicorn.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests", "tests.*"]
ignore_errors = true

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["pydantic_fixturegen", "tests"]
exclude = [".pfg-cache"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
  "SIM",
]
ignore = [
  "E203",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.hatch.build.targets.sdist]
include = [
  "pydantic_fixturegen",
  "tests",
  "docs",
  "README.md",
]

[tool.hatch.build.targets.wheel]
packages = ["pydantic_fixturegen"]
